Add utility getV2ClientNamesFromDefault
