+++ src/transforms/v2-to-v3/utils/get/index.ts
@@ -3,9 +3,9 @@ export * from "./getRequireVariableDeclaration";
 export * from "./getV2ClientIdentifiers";
 export * from "./getV2ClientIdThisExpressions";
 export * from "./getV2ClientNames";
+export * from "./getV2ClientNamesFromDefault";
 export * from "./getV2ClientTypeNames";
 export * from "./getV2DefaultModuleName";
-export * from "./getV2ServiceModuleNames";
 export * from "./getV2ServiceModulePath";
 export * from "./getV3ClientTypeName";
 export * from "./getV3ClientTypeNames";
